# Breaking Change: Removed extra handling of rules from `PageRequest`! (#152)
Rules and filters where not combinable (e.g. via `Filters.and(...)`) but from the API both are `QueryElements` and eventually produce a predicate for the `CursorPageRepositoryImpl`.

The separation was done for the serialization of the cursor, but this could also be solved in another way.

Here these two concepts are now merged into "filter", where rules are just another kind of filter now.

In order to clean up code, the previous methods for rules have already been removed.